requirements.txt
discord.py==2.1.1
pytz==2023.3